Fixed random text being appended glitch
Random text was being appended to the decrypted text. After a few hours of debugging, I figured out the problem. If the key was larger than the message then the program would break. I took out the check for resetting the key and ensured that the key wouldn't be bigger then the message by switching around the parameter calls in the encryptBlockBits method.